create index to export classes
- move sample code to sample.js
