+++ index.js
@@ -1,18 +1,4 @@
-import context from "./src/context.js";
-
-const Cli = context.$Cli();
-const Command = context.$Command();
-const console = context.console;
-const prompt = context.prompt;
-const question = context.$question();
-
-prompt.message = "cli-base";
-
-const cli = Cli.new([
-  Command.new("hello", async name => {
-    console.log(`hello ${name}`);
-  }, [ question.text("name") ])
-]);
-
-await cli.run();
-
+export { default as Cli } from "./src/Cli.js";
+export { default as Command } from "./src/Command.js";
+export { default as month } from "./src/month.js";
+export { default as question } from "./src/question.js";
